requirements.txt
anyio==3.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
fastapi==0.89.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.2
nordigen==1.3.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
python-dotenv==0.21.1
requests==2.28.2
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
